refactor(cleanup): extract shared checksum and scanning helpers
