+++ modules/gradle-plugin/src/main/kotlin/io/github/lmliam/microsmith/gradle/FilePathSeparator.kt
@@ -1,5 +1,7 @@
 package io.github.lmliam.microsmith.gradle
 
+import java.io.File
+
 internal object FilePathSeparator {
-    val value: String = System.getProperty("path.separator") ?: ":"
+    val value: String = File.pathSeparator
 }
